changes for 64-bit machines
SCCS-vsn: sys/sys/dirent.h 8.2
SCCS-vsn: sys/sys/stat.h 8.7

+++ usr/src/sys/sys/stat.h
@@ -9,7 +9,7 @@
  *
  * %sccs.include.redist.c%
  *
- *	@(#)stat.h	8.6 (Berkeley) %G%
+ *	@(#)stat.h	8.7 (Berkeley) %G%
  */
 
 #ifndef _SYS_STAT_H_
@@ -19,42 +19,42 @@
 
 #ifndef _POSIX_SOURCE
 struct ostat {
-	unsigned short	st_dev;		/* inode's device */
-	ino_t	st_ino;			/* inode's number */
-	mode_t	st_mode;		/* inode protection mode */
-	nlink_t	st_nlink;		/* number of hard links */
-	unsigned short	st_uid;		/* user ID of the file's owner */
-	unsigned short	st_gid;		/* group ID of the file's group */
-	unsigned short	st_rdev;	/* device type */
-	long	st_size;		/* file size, in bytes */
+	u_int16_t st_dev;		/* inode's device */
+	ino_t	  st_ino;		/* inode's number */
+	mode_t	  st_mode;		/* inode protection mode */
+	nlink_t	  st_nlink;		/* number of hard links */
+	u_int16_t st_uid;		/* user ID of the file's owner */
+	u_int16_t st_gid;		/* group ID of the file's group */
+	u_int16_t st_rdev;		/* device type */
+	int32_t	  st_size;		/* file size, in bytes */
 	struct	timespec st_atimespec;	/* time of last access */
 	struct	timespec st_mtimespec;	/* time of last data modification */
 	struct	timespec st_ctimespec;	/* time of last file status change */
-	long	st_blksize;		/* optimal blocksize for I/O */
-	long	st_blocks;		/* blocks allocated for file */
-	unsigned long	st_flags;	/* user defined flags for file */
-	unsigned long	st_gen;		/* file generation number */
+	int32_t	  st_blksize;		/* optimal blocksize for I/O */
+	int32_t	  st_blocks;		/* blocks allocated for file */
+	u_int32_t st_flags;		/* user defined flags for file */
+	u_int32_t st_gen;		/* file generation number */
 };
 #endif /* !_POSIX_SOURCE */
 
 struct stat {
-	dev_t	st_dev;			/* inode's device */
-	ino_t	st_ino;			/* inode's number */
-	mode_t	st_mode;		/* inode protection mode */
-	nlink_t	st_nlink;		/* number of hard links */
-	uid_t	st_uid;			/* user ID of the file's owner */
-	gid_t	st_gid;			/* group ID of the file's group */
-	dev_t	st_rdev;		/* device type */
+	dev_t	  st_dev;		/* inode's device */
+	ino_t	  st_ino;		/* inode's number */
+	mode_t	  st_mode;		/* inode protection mode */
+	nlink_t	  st_nlink;		/* number of hard links */
+	uid_t	  st_uid;		/* user ID of the file's owner */
+	gid_t	  st_gid;		/* group ID of the file's group */
+	dev_t	  st_rdev;		/* device type */
 	struct	timespec st_atimespec;	/* time of last access */
 	struct	timespec st_mtimespec;	/* time of last data modification */
 	struct	timespec st_ctimespec;	/* time of last file status change */
-	off_t	st_size;		/* file size, in bytes */
-	quad_t	st_blocks;		/* blocks allocated for file */
-	unsigned long	st_blksize;	/* optimal blocksize for I/O */
-	unsigned long	st_flags;	/* user defined flags for file */
-	unsigned long	st_gen;		/* file generation number */
-	long	st_lspare;
-	quad_t	st_qspare[2];
+	off_t	  st_size;		/* file size, in bytes */
+	quad_t	  st_blocks;		/* blocks allocated for file */
+	u_int32_t st_blksize;		/* optimal blocksize for I/O */
+	u_int32_t st_flags;		/* user defined flags for file */
+	u_int32_t st_gen;		/* file generation number */
+	int32_t	  st_lspare;
+	quad_t	  st_qspare[2];
 };
 #define st_atime st_atimespec.ts_sec
 #define st_mtime st_mtimespec.ts_sec
